Set correct locations for WebView's

+++ app/src/main/java/byui/cit360/employeereport/control/Weather.java
@@ -45,7 +45,7 @@ public class Weather {
     private void addTemperature() {
         try {
             JSONObject tempJson = (JSONObject) weatherInfo.get("main");
-            Double temp = (Double) tempJson.get("temp");
+            Integer temp = (Integer) tempJson.get("temp");
             long imp_temp = Math.round((temp - 270) * 9 / 5 + 32);
             weatherString += "Temp: " + imp_temp + "F\n";
         } catch (JSONException e) {
